refactor: bean 패키지 분리
