+++ nextstep-di/src/main/java/nextstep/di/factory/bean/ClassPathBeanDefinition.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory.scanner;
+package nextstep.di.factory.bean;
 
 public class ClassPathBeanDefinition implements BeanDefinition {
     private Class<?> clazz;
