+++ nextstep-di/src/main/java/nextstep/di/factory/scanner/Scanner.java
@@ -1,5 +1,7 @@
 package nextstep.di.factory.scanner;
 
+import nextstep.di.factory.bean.BeanDefinition;
+
 import java.util.Set;
 
 public interface Scanner {
