+++ nextstep-di/src/test/java/nextstep/di/factory/scanner/ConfigurationBeanScannerTest.java
@@ -1,5 +1,6 @@
 package nextstep.di.factory.scanner;
 
+import nextstep.di.factory.bean.BeanDefinition;
 import nextstep.stereotype.Controller;
 import nextstep.stereotype.Repository;
 import nextstep.stereotype.Service;
@@ -11,7 +12,6 @@ import java.util.Set;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
-import static org.junit.jupiter.api.Assertions.*;
 
 class ConfigurationBeanScannerTest {
     private static final Class[] AVAILABLE_ANNOTATIONS = {Controller.class, Service.class, Repository.class};
